Split CLI to fix blazor build issue
