+++ .vscode/launch.json
@@ -12,7 +12,7 @@
 			"outFiles": ["${workspaceRoot}/client/out/**/*.js"],
 			"preLaunchTask": {
 				"type": "npm",
-				"script": "build"
+				"script": "compile"
 			}
 		},
 		{
