+++ DevSkim-DotNet/Microsoft.DevSkim.Blazor/wwwroot/index.html
@@ -5,7 +5,8 @@
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
     <title>Microsoft.DevSkim.Blazor</title>
-    <base href="/DevSkim/" />
+    <!--<base href="/DevSkim/" />-->
+    <base href="/" />
     <link href="css/bootstrap/bootstrap.min.css" rel="stylesheet" />
     <link href="css/app.css" rel="stylesheet" />
     <link href="manifest.json" rel="manifest" />
