+++ DevSkim-DotNet/Microsoft.DevSkim.Blazor/wwwroot/service-worker.js
