+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Catalogue.cs
