+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Commands/AnalyzeCommand.cs
