+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Commands/CatalogueCommand.cs
