+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Commands/ICommand.cs
