+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Commands/PackCommand.cs
