+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Commands/VerifyCommand.cs
