+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/ErrorMessage.cs
