+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/ExceptionLogger.cs
