+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Globals.cs
