+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Tester.cs
