+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Verifier.cs
