+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Writers/DummyWriter.cs
