+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Writers/JsonWriter.cs
