+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Writers/SarifWriter.cs
