+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Writers/Writer.cs
