+++ DevSkim-DotNet/Microsoft.DevSkim.CLI.Commands/Writers/WriterFactory.cs
