+++ DevSkim-DotNet/Microsoft.DevSkim.CLI/Microsoft.DevSkim.CLI.csproj
@@ -27,13 +27,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Glob" Version="1.1.9" />
-    <PackageReference Include="Microsoft.CST.RecursiveExtractor" Version="1.1.6" />
-    <PackageReference Include="Microsoft.Extensions.CommandLineUtils" Version="1.1.1" />
-    <PackageReference Include="Sarif.Sdk" Version="2.4.12" />
-  </ItemGroup>
-
-  <ItemGroup>
+    <ProjectReference Include="..\Microsoft.DevSkim.CLI.Commands\Microsoft.DevSkim.CLI.Commands.csproj" />
     <ProjectReference Include="..\Microsoft.DevSkim\Microsoft.DevSkim.csproj" />
   </ItemGroup>
