+++ DevSkim-DotNet/Microsoft.DevSkim.sln
@@ -1,6 +1,6 @@
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29709.97
+# Visual Studio Version 17
+VisualStudioVersion = 17.0.32014.148
 MinimumVisualStudioVersion = 15.0.26124.0
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DevSkim", "Microsoft.DevSkim\Microsoft.DevSkim.csproj", "{6413D0EE-B522-47C4-A11A-A917426DD43F}"
 EndProject
@@ -14,9 +14,20 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DevSkim.VSExtensi
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DevSkim.VSExtension.Tests", "Microsoft.DevSkim.VSExtension.Tests\Microsoft.DevSkim.VSExtension.Tests.csproj", "{2E699F46-E217-4FB7-B375-C0522759BE82}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DevSkim.Blazor", "Microsoft.DevSkim.Blazor\Microsoft.DevSkim.Blazor.csproj", "{A9046024-40F3-4F7A-AAB5-FF13639A75B5}"
+Project("{D954291E-2A0B-460D-934E-DC6B0785DB48}") = "Microsoft.DevSkim.VSExtension.Shared", "Microsoft.DevSkim.VSExtension.Shared\Microsoft.DevSkim.VSExtension.Shared.shproj", "{7F009D24-6EB7-4185-A439-9260820095E3}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DevSkim.VSExtension.2022", "Microsoft.DevSkim.VSExtension.2022\Microsoft.DevSkim.VSExtension.2022.csproj", "{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Microsoft.DevSkim.CLI.Commands", "Microsoft.DevSkim.CLI.Commands\Microsoft.DevSkim.CLI.Commands.csproj", "{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Microsoft.DevSkim.Blazor", "Microsoft.DevSkim.Blazor\Microsoft.DevSkim.Blazor.csproj", "{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}"
 EndProject
 Global
+	GlobalSection(SharedMSBuildProjectFiles) = preSolution
+		Microsoft.DevSkim.VSExtension.Shared\Microsoft.DevSkim.VSExtension.Shared.projitems*{4ba5b90a-68d2-44ed-83cd-c96db44d9630}*SharedItemsImports = 4
+		Microsoft.DevSkim.VSExtension.Shared\Microsoft.DevSkim.VSExtension.Shared.projitems*{74de7137-f3ec-4c57-82f5-e66054a9659c}*SharedItemsImports = 4
+		Microsoft.DevSkim.VSExtension.Shared\Microsoft.DevSkim.VSExtension.Shared.projitems*{7f009d24-6eb7-4185-a439-9260820095e3}*SharedItemsImports = 13
+	EndGlobalSection
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64
@@ -98,18 +109,42 @@ Global
 		{2E699F46-E217-4FB7-B375-C0522759BE82}.Release|x64.Build.0 = Release|Any CPU
 		{2E699F46-E217-4FB7-B375-C0522759BE82}.Release|x86.ActiveCfg = Release|Any CPU
 		{2E699F46-E217-4FB7-B375-C0522759BE82}.Release|x86.Build.0 = Release|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Debug|x64.Build.0 = Debug|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Debug|x86.Build.0 = Debug|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Release|x64.ActiveCfg = Release|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Release|x64.Build.0 = Release|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Release|x86.ActiveCfg = Release|Any CPU
-		{A9046024-40F3-4F7A-AAB5-FF13639A75B5}.Release|x86.Build.0 = Release|Any CPU
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Debug|x64.Build.0 = Debug|Any CPU
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Debug|x86.ActiveCfg = Debug|x86
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Debug|x86.Build.0 = Debug|x86
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Release|x64.ActiveCfg = Release|Any CPU
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Release|x64.Build.0 = Release|Any CPU
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Release|x86.ActiveCfg = Release|x86
+		{4BA5B90A-68D2-44ED-83CD-C96DB44D9630}.Release|x86.Build.0 = Release|x86
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Debug|x64.Build.0 = Debug|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Debug|x86.Build.0 = Debug|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Release|x64.ActiveCfg = Release|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Release|x64.Build.0 = Release|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Release|x86.ActiveCfg = Release|Any CPU
+		{FB348F7D-1F09-4CBF-A0BB-C1AB7F6C505B}.Release|x86.Build.0 = Release|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Debug|x64.Build.0 = Debug|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Debug|x86.Build.0 = Debug|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Release|Any CPU.Build.0 = Release|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Release|x64.ActiveCfg = Release|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Release|x64.Build.0 = Release|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Release|x86.ActiveCfg = Release|Any CPU
+		{79E0D707-555D-43EB-BCE2-02FDCA0E2B67}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
