+++ DevSkim-DotNet/Microsoft.DevSkim/Microsoft.DevSkim.csproj
@@ -19,7 +19,7 @@
   </PropertyGroup>
   
   <ItemGroup>
-    <PackageReference Include="Microsoft.CST.OAT" Version="1.1.16" />
+    <PackageReference Include="Microsoft.CST.OAT" Version="1.1.19" />
     <PackageReference Include="System.Text.Json" Version="6.0.1" />
   </ItemGroup>
